fix: modal windows
