+++ src/components/AuthForm.tsx
@@ -38,28 +38,24 @@ const AuthForm = () => {
         <legend className="from-light-green-text to-light-blue-text mb-10 bg-gradient-to-r bg-clip-text text-center text-2xl font-medium text-transparent md:text-2xl">
           Вход
         </legend>
-        <label className="mb-3.5 flex flex-col items-start justify-start" htmlFor="email">
-          <span className="mb-2.5 text-xs uppercase text-[#2B2B2B] md:text-base">
-            АДРЕС ЭЛЕКТРОННОЙ ПОЧТЫ
-          </span>
-          <Input
-            error={errors.email ? true : false}
-            id="email"
-            name="email"
-            register={register}
-            type="email"
-          />
-        </label>
-        <label className="mb-2 flex flex-col items-start justify-start" htmlFor="password">
-          <span className="mb-2.5 text-xs uppercase text-[#2B2B2B] md:text-base">ПАРОЛЬ</span>
-          <Input
-            error={errors.password ? true : false}
-            id="password"
-            name="password"
-            register={register}
-            type="password"
-          />
-        </label>
+        <Input
+          error={errors.email ? true : false}
+          id="email"
+          label="АДРЕС ЭЛЕКТРОННОЙ ПОЧТЫ"
+          name="email"
+          placeholder=""
+          register={register}
+          type="email"
+        />
+        <Input
+          error={errors.password ? true : false}
+          id="password"
+          label="ПАРОЛЬ"
+          name="password"
+          placeholder=""
+          register={register}
+          type="password"
+        />
         <div className="mb-6 flex justify-between">
           {!isValid && (
             <p className="text-xs text-[#FF6F6F] md:text-base">
@@ -70,18 +66,21 @@ const AuthForm = () => {
             Забыли пароль?
           </a>
         </div>
-        <label className="mb-8 flex items-center" htmlFor="remember_password">
-          <Input
-            id="remember_password"
-            name="remember_password"
-            register={register}
-            type="checkbox"
-          />
-          <span className="ml-3 text-xs font-light text-[#2B2B2B] md:text-base">
-            Запомнить аккаунт
-          </span>
-        </label>
-        <Button disable={!isValid}>Войти</Button>
+        <Input
+          id="remember_password"
+          label="Запомнить аккаунт"
+          name="remember_password"
+          register={register}
+          type="checkbox"
+        />
+        {/* <Button disable={!isValid}>Войти</Button> */}
+        <Button
+          background="from-light-green to-light-blue bg-gradient-to-r"
+          disable={!isValid}
+          textColor="text-white"
+        >
+          {'Войти'}
+        </Button>
       </fieldset>
       <div className="mt-5 flex items-center justify-between">
         <Link className="text-xs font-light text-[#07836C] md:text-base" to="/register">
